refactor: move playSimilar action to playbackActions and streamline song processing
